refactor: hashTag를 component로 뻄
