+++ src/pages/profile-upload/index.tsx
@@ -5,7 +5,6 @@ import styles from '../../styles/UploadPage.module.css';
 import styleButton from '../../styles/Button.module.css';
 import BlankImage from '../../../public/icons/blank-image.svg';
 import Plus from '../../../public/icons/plus.svg';
-import Delete from '../../../public/icons/delete.svg';
 import Ect from '../../../public/icons/ETC.svg';
 import {
   IUploadFile,
@@ -15,6 +14,7 @@ import {
 } from '../../types/upload';
 import UploadVideo from '../../components/upload/UploadVideo';
 import DanceGenre from '@/components/upload/DanceGenre';
+import HashTag from '@/components/upload/HashTag';
 
 export default function ProfileUpload() {
   const [image, setImage] = useState<IUploadFile | null>(null);
@@ -131,43 +131,6 @@ export default function ProfileUpload() {
   };
 
   const nextId = useRef<number>(1);
-
-  //hashTag
-  const addHashTag = (e: any) => {
-    const hashTagItem = {
-      id: nextId.current,
-      hashTag: e.target.value.trim(),
-    };
-
-    const allowedCommand = ['Enter', 'NumpadEnter'];
-    if (!allowedCommand.includes(e.code)) return;
-
-    if (hashTagItem.hashTag == '') {
-      return;
-    } else {
-      if (hashTagList.length < 4) {
-        setHashTagList([...hashTagList, hashTagItem]);
-        setHashTag('');
-        nextId.current += 1;
-        console.log({ hashTagList });
-      } else {
-        setHashTag('');
-      }
-    }
-  };
-
-  const deleteHashTag = (id: number) => {
-    setHashTagList(hashTagList.filter(item => item.id !== id));
-  };
-
-  const KeyDownHadler = (e: React.KeyboardEvent) => {
-    if (e.code != 'Enter' && e.code != 'NumpadEnter') return;
-    e.preventDefault();
-  };
-
-  const handleChangeHashTag = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setHashTag(e.target.value);
-  };
 
   //Award,Date
   const addAward = () => {
@@ -306,44 +269,13 @@ export default function ProfileUpload() {
             </>
           )}
         </div>
-        <div className={styles.box}>
-          <div className={styles.maximum}>
-            <div className={styles.text}>나를 소개하는 해시태그</div>
-            <div className={styles.smallTexts}>
-              <div className={`${styles.smallText} ${styles.spacing}`}>
-                최대
-              </div>
-              <div className={`${styles.smallText} ${styles.pointText}`}>3</div>
-              <div className={styles.smallText}>개</div>
-            </div>
-          </div>
-          <input
-            className={`${styles.input} ${styles.long}`}
-            placeholder="빠른템포의, 허니제이같은 등의 키워드를 작성해보세요!"
-            type="text"
-            value={hashTag}
-            onChange={handleChangeHashTag}
-            onKeyUp={addHashTag}
-            onKeyDown={KeyDownHadler}
-          />
-        </div>
-        <div className={styles.hashTags}>
-          {hashTagList.length > 1 &&
-            hashTagList.slice(1, 4).map((item, idx) => (
-              <div key={idx} className={styles.hashTagBox}>
-                <div key={idx} className={styles.hashTag}>
-                  {'#' + item.hashTag}
-                </div>
-                <div
-                  onClick={() => deleteHashTag(item.id)}
-                  className={styles.deleteHashTag}
-                >
-                  <Delete />
-                </div>
-              </div>
-            ))}
-        </div>
-
+        <HashTag
+          hashTag={hashTag}
+          setHashTag={setHashTag}
+          hashTagList={hashTagList}
+          setHashTagList={setHashTagList}
+          title="나를 소개하는 해시태그"
+        />
         <div className={styles.box}>
           <div className={styles.text}>공연 및 활동경력</div>
           {awardList.map((item, idx) => (
