+++ src/styles/UploadPage.module.css
@@ -251,7 +251,7 @@
   display: flex;
   flex-direction: row;
   flex-wrap: wrap;
-  width: 339px;
+  width: 100%;
   margin: -60px 0 60px 0;
 }
